First Real Program to assist in My work!!!!
This is the first real crack at using Python with Excel to automate boring stuff. The code runs a routine task that is usually accomplished  2 to 3 times a week. The Raw data is refreshed/downloaded each time and formatted the same way. Next step wil be to set the proper locations for where the to find the raw data each time, then save in a speific location adding a date to the new file name when saved.  WOO HOO!!